OPENDJ-808 Implement a simple commons REST based HTTP connection handler
ConcurrentHashMap does not accept putting in null values, so I will put in the key itself (It is redundant, but this is not a problem here.)

HTTPConnectionHandler.java:
clientConnections changed from Map<ClientConnection, Void> to Map<ClientConnection, ClientConnection>.

CollectClientConnectionsFilter.java:
Consequence  of the change to HTTPConnectionHandler. 
Use the key to put in the value.