(CR-1535) Fixed QuickSetup.jnlp after adding the HTTP connection handler and all its jars
create-webstart-standalone.sh:
Iterate over the JARs in the lib directory in order to dynamically build the list of JARs.
Note: 'ant resolve' must be run prior to running this script when running this script from an svn checkout.
It is like this since Apache Ivy has been added and JARs have been removed from the 'lib' directory.